refactor: extract printSingleLambdaInvocation method to regroup common logic
