+++ packages/prettier-plugin-java/src/utils/index.ts
@@ -1,0 +1,2 @@
+export { default as printArgumentListWithBraces } from "./printArgumentListWithBraces";
+export { default as printSingleLambdaInvocation } from "./printSingleLambdaInvocation";
