+++ packages/prettier-plugin-java/src/utils/printArgumentListWithBraces.ts
@@ -1,39 +1,28 @@
 import { ArgumentListCstNode, IToken } from "java-parser";
 import { builders } from "prettier/doc";
-import {
-  isArgumentListSingleLambda,
-  isSingleArgumentLambdaExpressionWithBlock
-} from "./expressions-utils";
-import { printTokenWithComments } from "../printers/comments/format-comments";
-import { concat, dedent, indent } from "../printers/prettier-builder";
+import { isArgumentListSingleLambda } from "./expressions-utils";
 import { putIntoBraces } from "../printers/printer-utils";
+import printSingleLambdaInvocation from "./printSingleLambdaInvocation";
 
-const { softline, ifBreak } = builders;
+const { softline } = builders;
 
-export function printArgumentListWithBraces(
+export default function printArgumentListWithBraces(
   argumentListCtx: ArgumentListCstNode[] | undefined,
   rBrace: IToken,
   lBrace: IToken
 ) {
-  const lambdaParametersGroupId = Symbol("lambdaParameters");
-  const argumentList = this.visit(argumentListCtx, {
-    lambdaParametersGroupId,
-    isInsideMethodInvocationSuffix: true
-  });
   const isSingleLambda = isArgumentListSingleLambda(argumentListCtx);
-
   if (isSingleLambda) {
-    const formattedRBrace = isSingleArgumentLambdaExpressionWithBlock(
-      argumentListCtx
-    )
-      ? ifBreak(
-          indent(concat([softline, rBrace])),
-          printTokenWithComments(rBrace),
-          { groupId: lambdaParametersGroupId }
-        )
-      : indent(concat([softline, rBrace]));
-    return dedent(putIntoBraces(argumentList, "", lBrace, formattedRBrace));
+    return printSingleLambdaInvocation.call(
+      this,
+      argumentListCtx,
+      rBrace,
+      lBrace
+    );
   }
 
+  const argumentList = this.visit(argumentListCtx, {
+    isInsideMethodInvocationSuffix: true
+  });
   return putIntoBraces(argumentList, softline, lBrace, rBrace);
 }
